Use context + hook for explorer state
